+++ packages/teams-js/test/internal/videoFrameTick.spec.ts
